Use size_t in array AM interface instead of int.

+++ include/array_bulk.h
@@ -16,10 +16,10 @@ void gather_init();
 /* AM-based scatter-gather primitives (for internal use in ti_array.c only) */
 
 /* rectangular scatter/gather */
-void get_array(void *pack_method, void *copy_desc, int copy_desc_size, 
+void get_array(void *pack_method, void *copy_desc, size_t copy_desc_size, 
                int tgt_box, void *buffer, int atomicelements);
 void put_array(void *unpack_method, void *copy_desc, int copy_desc_size,
-               void *array_data, int array_data_size, int tgt_box);
+               void *array_data, size_t array_data_size, int tgt_box);
 
 /* sparse scatter/gather - transfer data back & forth between a 
  * local, packed data list and the memory space of a remote proc
@@ -28,9 +28,9 @@ void put_array(void *unpack_method, void *copy_desc, int copy_desc_size,
  *           sizeof(remote_addr_list) == num_elem * sizeof(void*)
  */
 void sparse_scatter(void **remote_addr_list, void *src_data_list, 
-                    int remote_box, int num_elem, int elem_sz, int atomic_elements);
+                    int remote_box, size_t num_elem, size_t elem_sz, int atomic_elements);
 void sparse_gather(void *tgt_data_list, void **remote_addr_list, 
-                   int remote_box, int num_elem, int elem_sz, int atomic_elements);
+                   int remote_box, size_t num_elem, size_t elem_sz, int atomic_elements);
 
 
 /* handler declarations */
